+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 Flutter Project Setup CLI Changelog
 
+## 0.0.3 - 03-09-2024
+
+### Fixed
+- Resolved an issue where CLI would fail to generate required files when outside the project directory
+
+### Added
+- Added tests for mason brick generations
+
 ## 0.0.2 - 21-08-2024
 
 - Added tests for the create command